Fix another audio offset bug
The audio sometimes would start playing for a short moment at the beginning of the video even when it should be delayed
Fix by filling start of audio stream with silence.
While this fixes the bug I'm  not sure I'm doing everything exactly right. The ffmpeg api is a dark place.